backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.6
jiter==0.6.1
openai==1.51.2
pydantic==2.6.3
pydantic_core==2.16.3
python-decouple==3.8
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.5
typing_extensions==4.11.0
uvicorn==0.28.0
